# Optimize LayerNorm performance on CPU both forward and backward