@mingfeima 2020-03-31
Currently, forward path of `LayerNorm` is partially vectorized, backward is neither parallelized nor vectorized.

Results on Xeon 6248, 2x20 cores @ 2.5GHz. Use [benchmark1](https://github.com/mingfeima/op_bench-py) to reproduce, `./run.sh layernorm.py`

Input size: [128,128,1024], Output size: [1024], Unit: ms per iteration (the lower the better)

Input Size | original | this pr | speedup
-- | -- | -- | --
single socket inference | 2.16 | 1.89 | 1.14
single core inference | 43.44 | 24.59 | 1.77
single socket training | 42.78 | 6.82 | 6.27


Fine tuning result on [gpt2](https://github.com/huggingface/transformers/tree/master/examples#gpt-2gpt-and-causal-language-modeling) language modeling fine tuning on WikiTest2 dataset:
```bash
###  before
| 591/591 [46:11<00:00,  4.69s/it]
### after
| 591/591 [31:09<00:00,  3.16s/it]
```