@glaringlee 2020-08-03
@mingfeima 
Please rebase the code, the macos CI failure will go away.