@glaringlee 2020-08-10
> Hi @glaringlee i am seeing some upstream breakage during the rebase:
> 
> ```shell
> fatal: reference is not a tree: f015d698006c4a11be15b1ebb75b3b9bb317b914
> Unable to checkout 'f015d698006c4a11be15b1ebb75b3b9bb317b914' in submodule path 'third_party/tensorpipe'
> ```
> 
> The last 7 commits from tensorpipe repo mismatch pytorch third_party:
> https://github.com/pytorch/tensorpipe/commits/f015d698006c4a11be15b1ebb75b3b9bb317b914
> https://github.com/pytorch/tensorpipe/commits/master
> 
> Any idea how to solve this?

@mingfeima please try again, we just fixed it.